auth forms optimized for DX
